configured unit of each buy

+++ apps/reports/models.py
@@ -5,7 +5,7 @@ class Reports(models.Model):
     product = models.CharField(max_length=255)
     category = models.CharField(max_length=255)
     quantity_itens = models.IntegerField()
-    stock = models.IntegerField()
+    stock = models.DecimalField(max_digits=10, decimal_places=2)
     sale = models.DecimalField(max_digits=10, decimal_places=2)
     payment = models.CharField(max_length=20)
     data = models.DateField(null=True, blank=True)
